requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cloudscraper==1.2.71
curl_cffi==0.14.0
feedparser==6.0.12
greenlet==3.3.2
h11==0.16.0
idna==3.11
outcome==1.3.0.post0
packaging==26.0
playwright==1.58.0
pycparser==3.0
pyee==13.0.1
pyparsing==3.3.2
PySocks==1.7.1
python-dotenv==1.2.2
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.41.0
setuptools==82.0.1
sgmllib3k==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
undetected-chromedriver==3.5.5
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2